move student db specs to new file
